profile filtration

+++ apps/categories/usecases/category_usecase.py
@@ -1,4 +1,5 @@
 from apps.categories.repositories.category_repository import CategoryRepository
+from common.exceptions import NotFoundException, PermissionDeniedError
 from rest_framework.exceptions import NotFound
 
 
@@ -9,19 +10,28 @@ class CategoryUseCase:
     def list(self, level=None, search=None):
         return self.repo.get_all(level=level, search=search)
 
+    def list_by_user(self, user, search=None):
+        return self.repo.get_by_creator(user, search=search)
+
     def get(self, pk):
         obj = self.repo.get_by_id(pk)
         if not obj:
             raise NotFound("Category not found")
         return obj
 
-    def create(self, **data):
+    def create(self, user=None, **data):
+        if user is not None:
+            data["creator_id"] = user
         return self.repo.create(**data)
 
-    def update(self, pk, **data):
+    def update(self, pk, user, **data):
         obj = self.get(pk)
+        if obj.creator_id != user:
+            raise PermissionDeniedError("Только создатель может редактировать запись")
         return self.repo.update(obj, **data)
 
-    def delete(self, pk):
+    def delete(self, pk, user):
         obj = self.get(pk)
+        if obj.creator_id != user:
+            raise PermissionDeniedError("Только создатель может удалить запись")
         return self.repo.delete(obj)
